Rewrite sflag handling
In this version the -s and the -d flags are exclusive and cannot
both be specified. The intended usage model is as follows:

  tcplay -m disk -s /dev/sda
  kpartx -a /dev/mapper/disk

The /dev/mapper/disk represents the raw disk in a decrypted state
and kpartx is used to parse the partition table and create
additional mappings for each partition.

Signed-off-by: Mike Baker <mbm@openwrt.org>